Removed some now unnecessary casts! Fixed string name for oredict missing tooltip in manual!

+++ src/main/java/li/cil/tis3d/common/module/ModuleExecution.java
@@ -343,10 +343,10 @@ public final class ModuleExecution extends AbstractModuleRotatable {
         GlStateManager.color(1f, 1f, 1f, 1f);
 
         // Draw register info on top.
-        final String accLast = String.format("ACC:%4X LAST:%s", (short) machineState.acc, machineState.last.map(Enum::name).orElse("NONE"));
+        final String accLast = String.format("ACC:%4X LAST:%s", machineState.acc, machineState.last.map(Enum::name).orElse("NONE"));
         FontRendererAPI.drawString(accLast);
         GlStateManager.translate(0, FontRendererAPI.getCharHeight() + 4, 0);
-        final String bakState = String.format("BAK:%4X MODE:%s", (short) machineState.bak, state.name());
+        final String bakState = String.format("BAK:%4X MODE:%s", machineState.bak, state.name());
         FontRendererAPI.drawString(bakState);
         GlStateManager.translate(0, FontRendererAPI.getCharHeight() + 4, 0);
         drawLine(1);
